feat+perf(download_points): parallel point fetching (#564)
* feat+perf(download_points): parallel point fetching

Precursor to fancy indexing.

* test: check download_points has same result

* feat: add scattered_points function

* fix: get_root (deprecated) was broken

* fix: graphene scattered_points works

* test: add test for precomputed scattered_points

* docs: show how to use scattered_points

* perf: remove redundant requests

* fix: really eliminate redundant requests

+++ cloudvolume/datasource/precomputed/metadata.py
@@ -632,7 +632,7 @@ Hops:
   def point_to_mip(self, pt, mip, to_mip):
     pt = Vec(*pt)
     downsample_ratio = self.resolution(mip).astype(np.float32) / self.resolution(to_mip).astype(np.float32)
-    return np.floor(pt * downsample_ratio)
+    return np.floor(pt * downsample_ratio).astype(np.int64)
 
   def bbox_to_mip(self, bbox, mip, to_mip):
     """Convert bbox or slices from one mip level to another."""
